Create BULGAR_ALARM
A burglar alarm will be activated if an unauthorized person is detected by a Window Sensor(variable WS) or a Motion Detector(Variable MD). This has to do only when the Burglar Alarm System(variable SYS) is activated. The output of the Burglar alarm system is denoted by variable ALARM